+++ srcs/vec3-operations.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   vec3-operations.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hmontoya <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 14:52:12 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/16 14:52:17 by hmontoya         ###   ########.fr       */
+/*   Updated: 2024/12/17 11:39:45 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vectors.h"
 
-t_vec3 vec3(double x, double y, double z)
+t_vec3 vec3(float x, float y, float z)
 {
 	t_vec3 v3;
 	v3.x = x;
@@ -57,4 +57,4 @@ t_vec3		rest_v3(t_vec3 v1, t_vec3 v2)
 	new.z = v1.z - v2.z;
 	return (new);
 }
-//return ((double)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
+//return ((float)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
